simplify crypto_sign_sr25519

+++ app/src/crypto.h
@@ -49,7 +49,7 @@ zxerr_t crypto_fillAddress(key_kind_e addressKind, uint8_t *buffer, uint16_t buf
 
 void copy_sr25519_signdata(uint8_t *buffer);
 
-zxerr_t crypto_sign_sr25519(const uint8_t *message, uint16_t messageLen);
+zxerr_t crypto_sign_sr25519(const uint8_t *message, size_t messageLen);
 
 zxerr_t crypto_sign_ed25519(uint8_t *signature, uint16_t signatureMaxlen,
                             const uint8_t *message, uint16_t messageLen);
